==> notebooks/01_tb_scrape_directory.ipynb <==
# 01. Timebanks: Directory Scraper
> Author: [Dawn Graham](https://dawngraham.github.io/)

Scrape all pages of TimeBanks.org directory to get initial listing of all timebanks.

## Import libraries

import pandas as pd
import requests
import time
import unicodedata
import regex as re
from bs4 import BeautifulSoup

## Get slugs

tb_slugs = []

counter = 0
print('Getting timebank slugs... Directory page # ')

# Cycle through all pages in directory
for page in range(16):
    url = f'http://community.timebanks.org/directory?page={page}&js=1&order=title_1&sort=asc'
    res = requests.get(url)
    soup = BeautifulSoup(res.content, 'lxml')
    table = soup.find('table')
    
    for row in table.find('tbody').find_all('tr'):
        tb_slug = {}
        tb_slug['name'] = row.find('a').text.strip()
        tb_slug['slug'] = row.find('a').attrs['href'].strip('/')
        tb_slugs.append(tb_slug)
        
    print(counter+1, end=' ')
    
    time.sleep(1)
    counter += 1

pd.DataFrame(tb_slugs).head()

## Get timebank details

timebanks = []
counter = 0
total_timebanks = pd.DataFrame(tb_slugs).shape[0]

print(f'Getting {total_timebanks} timebank details... ')
for tb in range(len(tb_slugs)):
    url = f"http://community.timebanks.org/{tb_slugs[tb]['slug']}"
    res = requests.get(url)
    soup = BeautifulSoup(res.content, 'lxml')

    timebank = {}
    timebank['name'] = tb_slugs[tb]['name']
    timebank['slug'] = tb_slugs[tb]['slug']
    timebank['focus'] = soup.find('div', {'class': 'views-field-name'}).span.text.strip()
    try:
        timebank['sponsor'] = soup.find('div', {'class': 'views-field-field-tb-sponsor-value'}).span.text.strip()
    except:
        pass
    timebank['url'] = soup.find('div', {'class': 'views-field-markup-1'}).a['href'].strip()
    try:
        timebank['url_2'] = soup.find('div', {'class': 'views-field-field-custom-url-url'}).a.text.strip()
    except:
        pass
    timebank['address'] = soup.find('div', {'class': 'views-field-street'}).span.get_text(', ').strip()
    timebank['postal'] = soup.find('div', {'class': 'views-field-postal-code'}).span.text.strip()
    timebank['country'] = soup.find('div', {'class': 'views-field-country'}).span.text.strip()
    try:
        timebank['phone'] = soup.find('div', {'class': 'views-field-phone'}).span.text.strip()
    except:
        pass
    timebank['notes'] = soup.find('div', {'class': 'views-field-body'}).div.text.replace('\xa0',' ').replace('\n', ' ')
    timebank['timebank'] = re.sub('http://','', timebank['url']).replace('.timebanks.org', '')

    # Get social media links
    try:
        url = f"{timebank['url']}"
        res = requests.get(url)
        soup = BeautifulSoup(res.content, 'lxml')

        try:
            timebank['facebook'] = soup.find('div', {'class': 'fb-like'}).get('data-href').strip()
        except:
            pass
        try:
            timebank['twitter'] = soup.find('a', {'class': 'twitter-follow-button'}).get('href').strip()
        except:
            pass
    except:
        pass
    
    # Get mission statement
    try:
        url = f"{timebank['url']}/mission"
        res = requests.get(url)
        soup = BeautifulSoup(res.content, 'lxml')

        timebank['mission'] = soup.find('div', {'class': 'page-content'}).text.strip().replace('\xa0',' ').replace('\n', ' ').replace('\r', '')
    except:
        pass
    
    timebanks.append(timebank)
        
    if (counter+1) % 5 == 0:
        print(counter+1, end=' ')
    
    time.sleep(1)
    counter += 1
    if counter == total_timebanks:
        print('Done.')

# Save to dataframe
timebanks = pd.DataFrame(timebanks)

# Do not include demo or test-launch in directory
timebanks = timebanks[timebanks['timebank'] != 'democw3']
timebanks = timebanks[timebanks['timebank'] != 'test-launch-2018-07-28-2058']

# Save to .csv
timebanks.to_csv('../data/directory.csv', index=False)

# Preview
timebanks.head()
