==> notebooks/02_tb_daily_scraper.ipynb <==
# 02. Timebanks: Daily Scraper
> Author: [Dawn Graham](https://dawngraham.github.io/)

Get updates from TimeBanks.org at the beginning of each day:
- Numbers for exchanges, hours, members, offers, requests, last exchange
- Offers, requests, and talents by category
- All offer and request listings

On first initial run, set `header=True` to establish header in all .csv files. After, set `header=False` to append to existing files.

## Import Libraries

import pandas as pd
import numpy as np
import requests
import time
import unicodedata
import regex as re
from bs4 import BeautifulSoup

## Read in timebank slugs and url's

timebanks = pd.read_csv('../data/directory.csv', usecols=['slug', 'url'])
total_timebanks = timebanks.shape[0]
headers = {'User-agent': 'TimeBank update collector'}

## Get updated details

After initial collection, the code below can be run to get updated numbers for exchanges, hours, last_exchange, members, offers, and requests. This information can be used to get a sense of activity over time.

updates = []
counter = 0

print(f'Getting {total_timebanks} timebank details... ')
for timebank in range(len(timebanks)):
    url = f"http://community.timebanks.org/{timebanks['slug'][timebank]}"
    res = requests.get(url, headers=headers)
    soup = BeautifulSoup(res.content, 'lxml')

    update = {}
    update['members'] = soup.find('div', {'class': 'views-field-field-num-users-value'}).span.text.strip().replace(',', '')
    update['last_exchange'] = soup.find('div', {'class': 'views-field-field-last-exchange-value'}).span.text.strip()
    update['offers'] = soup.find('div', {'class': 'views-field-field-active-offers-value'}).span.text.strip().replace(',', '')
    update['requests'] = soup.find('div', {'class': 'views-field-field-active-requests-value'}).span.text.strip().replace(',', '')
    update['exchanges'] = soup.find('div', {'class': 'views-field-field-num-exchanges-value'}).span.text.strip().replace(',', '')
    update['hours'] = soup.find('div', {'class': 'views-field-field-hours-exchanged-value'}).span.text.strip().replace(',', '')
    update['timebank'] = re.sub('http://','', timebanks['url'][timebank]).replace('.timebanks.org', '')
    
    updates.append(update)
        
    if (counter+1) % 5 == 0:
        print(counter+1, end=' ')
    
    time.sleep(1)
    counter += 1
    if counter == total_timebanks:
        print('Done.')

# Convert to dataframe
updates = pd.DataFrame(updates)

# Add timestamp
filetime = time.strftime("%y%m%d_%H%M%S", time.localtime())
updates['timestamp'] = pd.to_datetime(filetime, format='%y%m%d_%H%M%S')

# Append results to .csv file
with open('../data/updates.csv', 'a') as file:
    updates.to_csv(file, header=False, index=False)

# Preview
print(updates.shape)
updates.head(10)

## Get offers by category

offers = []

counter = 0

print(f'Getting offers from {total_timebanks} timebanks... ')

# Iterate through all timebank directories
for timebank in range(len(timebanks)):
    url = f"{timebanks['url'][timebank]}/offers"
    res = requests.get(url, headers=headers)
    soup = BeautifulSoup(res.content, 'lxml')

    timebank_name = re.sub('http://','', timebanks['url'][timebank]).replace('.timebanks.org', '')
    
    try:
        # Get all parent categories
        for panel in soup.findAll('div', {'class': 'cw-panel'}):
            category = {}
            cat = panel.a
            category['count_offers'] = int(cat.span.text)
            cat.span.clear()
            parent_id = int(cat.get('href').strip('/ads?type=1&amp;cat='))
            category['cat_id'] = parent_id
            category['cat_parent'] = 'is_parent'
            category['category'] = cat.get_text()
            category['timebank'] = timebank_name
            offers.append(category)

            # Get all child categories
            for li in panel.findAll('li'):
                try:
                    category = {}
                    category['cat_id'] = int(li.a.get('href').strip('/ads?type=1&amp;cat='))
                    category['cat_parent'] = parent_id
                    category['category'] = li.a.get_text()
                    category['count_offers'] = int(li.span.text)
                    category['timebank'] = timebank_name
                    offers.append(category)
                except:
                    pass
    except:
        pass

    if (counter+1) % 5 == 0:
        print(counter+1, end=' ')
    
    time.sleep(1)
    counter += 1
    if counter == total_timebanks:
        print('Done.')

# Save to dateframe and drop duplicates
offers = pd.DataFrame(offers)
offers.drop_duplicates(inplace=True)
    
# Add timestamp
filetime = time.strftime("%y%m%d_%H%M%S", time.localtime())
offers['timestamp'] = pd.to_datetime(filetime, format='%y%m%d_%H%M%S')

# Append results to .csv file
with open('../data/offers.csv', 'a') as file:
    offers.to_csv(file, header=False, index=False)

# Preview
print(offers.shape)
offers.head(10)

## Get requests by category

reqs = []

counter = 0

print(f'Getting requests from {total_timebanks} timebanks... ')

# Iterate through all timebank directories
for timebank in range(len(timebanks)):
    url = f"{timebanks['url'][timebank]}/requests"
    res = requests.get(url, headers=headers)
    soup = BeautifulSoup(res.content, 'lxml')

    timebank_name = re.sub('http://','', timebanks['url'][timebank]).replace('.timebanks.org', '')
    
    try:
        # Get all parent categories
        for panel in soup.findAll('div', {'class': 'cw-panel'}):
            category = {}
            cat = panel.a
            category['count_requests'] = int(cat.span.text)
            cat.span.clear()
            parent_id = int(cat.get('href').strip('/ads?type=2&amp;cat='))
            category['cat_id'] = parent_id
            category['cat_parent'] = 'is_parent'
            category['category'] = cat.get_text()
            category['timebank'] = timebank_name
            reqs.append(category)

            # Get all child categories
            for li in panel.findAll('li'):
                try:
                    category = {}
                    category['cat_id'] = int(li.a.get('href').strip('/ads?type=2&amp;cat='))
                    category['cat_parent'] = parent_id
                    category['category'] = li.a.get_text()
                    category['count_requests'] = int(li.span.text)
                    category['timebank'] = timebank_name
                    reqs.append(category)
                except:
                    pass
    except:
        pass

    if (counter+1) % 5 == 0:
        print(counter+1, end=' ')
    
    time.sleep(1)
    counter += 1
    if counter == total_timebanks:
        print('Done.')

# Save to dateframe and drop duplicates
reqs = pd.DataFrame(reqs)
reqs.drop_duplicates(inplace=True)
    
# Add timestamp
filetime = time.strftime("%y%m%d_%H%M%S", time.localtime())
reqs['timestamp'] = pd.to_datetime(filetime, format='%y%m%d_%H%M%S')

# Append results to .csv file
with open('../data/requests.csv', 'a') as file:
    reqs.to_csv(file, header=False, index=False)

# Preview
print(reqs.shape)
reqs.head(10)

## Get talents by category
Note: Number of Timebankers' talents on timebank homepage does not correspond to number when you click on category link.

talents = []

counter = 0

print(f'Getting talents from {total_timebanks} timebanks... ')

# Iterate through all timebank directories
for timebank in range(len(timebanks)):
    url = f"{timebanks['url'][timebank]}/directory"
    res = requests.get(url, headers=headers)
    soup = BeautifulSoup(res.content, 'lxml')

    timebank_name = re.sub('http://','', timebanks['url'][timebank]).replace('.timebanks.org', '')
    
    try:
        # Get all parent categories
        for panel in soup.findAll('div', {'class': 'cw-panel'}):
            category = {}
            cat = panel.a
            category['count_talent'] = int(cat.span.text)
            cat.span.clear()
            parent_id = int(cat.get('href').strip('/directory?category='))
            category['cat_id'] = parent_id
            category['cat_parent'] = 'is_parent'
            category['category'] = cat.get_text()
            category['timebank'] = timebank_name
            talents.append(category)

            # Get all child categories
            for li in panel.findAll('li'):
                try:
                    category = {}
                    category['cat_id'] = int(li.a.get('href').strip('/directory?category='))
                    category['cat_parent'] = parent_id
                    category['category'] = li.a.get_text()
                    category['count_talent'] = int(li.span.text)
                    category['timebank'] = timebank_name
                    talents.append(category)
                except:
                    pass
    except:
        pass

    if (counter+1) % 5 == 0:
        print(counter+1, end=' ')
    
    time.sleep(1)
    counter += 1
    if counter == total_timebanks:
        print('Done.')

# Save to dateframe and drop duplicates
talents = pd.DataFrame(talents)
talents.drop_duplicates(inplace=True)
    
# Add timestamp
filetime = time.strftime("%y%m%d_%H%M%S", time.localtime())
talents['timestamp'] = pd.to_datetime(filetime, format='%y%m%d_%H%M%S')

# Append results to .csv file
with open('../data/talents.csv', 'a') as file:
    talents.to_csv(file, header=False, index=False)

# Preview
print(talents.shape)
talents.head(10)

## Get all offer & request listings
**Order to capture listings:**  
- For both listings types (1 = offers, 2 = requests)
    - get all timebanks
        - get all pages for each timebank
            - get all listings on each page

listings = []

print(f'Getting listings from {total_timebanks} timebanks... ')

# Get type 1 (offers) and type 2 (requests)
for i in range(1, 3):
    
    print(f'Getting type {i}... ', end=' ')
    counter = 0
    
    # Iterate through all timebank directories
    for timebank in range(len(timebanks)):
        url = f"{timebanks['url'][timebank]}/ads?so=desc&o=updated&limit=100&type={i}"
        res = requests.get(url, headers=headers)
        soup = BeautifulSoup(res.content, 'lxml')
        
        try:
            # Get total number of listings
            get_total = soup.find('div', {'class': 'pager-state'}).text
            total_listings = re.sub(r'(Showing \d+ - \d+ of )', '', get_total)
            try:
                total_listings = (int(total_listings))
            except:
                total_listings = 0

            # Get number of pages needed to get all listings
            pages = total_listings//100 + 1

            # Iterate through pages
            for page in range(pages):

                url = f"{timebanks['url'][timebank]}/ads?so=desc&o=updated&limit=100&type={i}&offset={page * 100}"
                res = requests.get(url)
                soup = BeautifulSoup(res.content, 'lxml')

                timebank_name = re.sub('http://','', timebanks['url'][timebank]).replace('.timebanks.org', '')

                for row in soup.findAll('div', {'class': 'media'}):

                    listing = {}

                    # Get listing id
                    listing['listing_id'] = int(row.find('input', {'class': ' selection-id'}).get('value'))

                    # Add all parent and child categories for a listing to `cat` set
                    cat = set()

                    for parent in row.findAll('a', {'class': 'parent'}):
                        cat.add(int(parent.get('href').strip('/ads?cat=')))

                    for child in row.findAll('a', {'class': 'child'}):
                        cat.add(int(child.get('href').strip('/ads?cat=')))

                    listing['listing_cat'] = cat

                    # Get listing description
                    listing['description'] = row.h4.text.strip()

                    # Add timebank name
                    listing['timebank'] = timebank_name

                    # Add listing type
                    if i == 1:
                        listing['type'] = 'offer'
                    else:
                        listing['type'] = 'request'

                    listings.append(listing)
                    
        except:
            pass
                
        if (counter+1) % 5 == 0:
            print(counter+1, end=' ')

        time.sleep(1)
        counter += 1
        if counter == total_timebanks:
            print('Done.')
        
# Convert to dataframe
listings = pd.DataFrame(listings)

# Add timestamp
filetime = time.strftime("%y%m%d_%H%M%S", time.localtime())
listings['timestamp'] = pd.to_datetime(filetime, format='%y%m%d_%H%M%S')

# Append results to .csv file
with open('../data/listings.csv', 'a') as file:
    listings.to_csv(file, header=False, index=False)

# Preview
print(listings.shape)
listings.head(10)
