==> notebooks/03_tb_cleaning_engineering.ipynb <==
# 03. Timebanks: Cleaning & Feature Engineering
> Author: [Dawn Graham](https://dawngraham.github.io/)

Clean, create features, and combine datasets into single .csv for modeling.

## Import libraries

import pandas as pd
import numpy as np
import regex as re
import time

import warnings
warnings.simplefilter(action='ignore')

from datetime import timedelta
from vaderSentiment.vaderSentiment import SentimentIntensityAnalyzer

## Functions

# Import data
def open_data(data_set):
    data_set = pd.read_csv(f'../data/{str(data_set)}.csv')

    try:
        # Set to timeseries
        data_set['timestamp'] = pd.to_datetime(data_set['timestamp'])
    except:
        pass
    
    return data_set

## Cleaning & Feature Engineering

### Create `df` starting with `directory.csv`

df = open_data('directory')
df.head(3)

# Drop `focus` column - no timebanks have value for this
df.drop(columns='focus', inplace=True)

# Drop unnecessary columns
df.drop(columns=['name', 'slug', 'url'], inplace=True)

### Sentiment analysis
Reference: [Simplifying Sentiment Analysis using VADER in Python (on Social Media Text)](https://medium.com/analytics-vidhya/simplifying-social-media-sentiment-analysis-using-vader-in-python-f9e6ec6fc52f)

# Get compound sentiment score for mission statements and notes
analyser = SentimentIntensityAnalyzer()

df['mission_sentiment'] = ''
df['notes_sentiment'] = ''

for row in range(len(df)):
    try:
        df['mission_sentiment'][row] = analyser.polarity_scores(df['mission'][row])['compound']
    except:
        df['mission_sentiment'][row] = 0
    try:
        df['notes_sentiment'][row] = analyser.polarity_scores(df['notes'][row])['compound']
    except:
        df['notes_sentiment'][row] = 0

### Change to numerical data

# Get word count for mission statement and notes
df['mission_word_count'] = ''
df['notes_word_count'] = ''

for row in range(len(df)):
    try:
        df['mission_word_count'][row] = len(df['mission'][row].split())
    except:
        df['mission_word_count'][row] = 0
    try:
        df['notes_word_count'][row] = len(df['notes'][row].split())
    except:
        df['notes_word_count'][row] = 0

# Drop `mission` and `notes` columns
df.drop(columns=['mission', 'notes'], inplace=True)

# Change 5555555555 phone number to NaN
df['phone'][6] = np.nan

# Set columns to 1 if timebank has feature, 0 if not
df[['phone', 'sponsor', 'url_2', 'facebook', 'twitter']] = df[['phone', 'sponsor', 'url_2', 'facebook', 'twitter']].notnull().astype(int)

# Reorder columns, including only timebank and numerical
df = df[['timebank', 'sponsor', 'phone', 'url_2', 'facebook', 'twitter', 'mission_word_count', 'mission_sentiment', 'notes_word_count', 'notes_sentiment']]

df.head()

### Add info from `updates.csv`

updates = open_data('updates')
updates.head()

# Convert `last_exchange` to date, time
updates['last_exchange_time'] = ''

for i in range(len(updates)):
    try:
        year = int(re.search('(\d+)\syear', updates.iloc[i]['last_exchange']).group(1))
    except:
        year = 0
        
    try:
        week = int(re.search('(\d+)\sweek', updates.iloc[i]['last_exchange']).group(1))
    except:
        week = 0
        
    try:
        day = int(re.search('(\d+)\sday', updates.iloc[i]['last_exchange']).group(1))
    except:
        day = 0
        
    try:
        hour = int(re.search('(\d+)\shour', updates.iloc[i]['last_exchange']).group(1))
    except:
        hour = 0
        
    try:
        minute = int(re.search('(\d+)\smin', updates.iloc[i]['last_exchange']).group(1))
    except:
        minute = 0
        
    try:
        sec = int(re.search('(\d+)\ssec', updates.iloc[i]['last_exchange']).group(1))
    except:
        sec = 0
    
    start = updates['timestamp'][i]
    end = timedelta(weeks=year*365 + week, days=day, hours=hour, minutes=minute, seconds=sec)
    
    updates['last_exchange_time'][i] = start - end

# Change to datetime
updates['last_exchange_time'] = pd.to_datetime(updates['last_exchange_time'])

# Set `last_exchange_time` to nan if start up with no exchanges
updates.loc[updates['exchanges'] == 'none (start up)', 'last_exchange_time'] = np.nan

# Replace `none (start up)` with 0 for all columns
updates.replace('none (start up)', 0, inplace=True)

# Change to `int`
updates['exchanges'] = updates['exchanges'].astype(int)
updates['hours'] = updates['hours'].astype(int)

updates.head()

**Daily Averages**

# Get total number of days covered in dataset
total_days = (updates['timestamp'].max() - updates['timestamp'].min()).round('D').days

# Get average number of daily exchanges and hours
df['avg_daily_exchanges'] = ''
df['avg_daily_hours'] = ''
df['hours_per_exchange'] = ''

for row in range(len(df)):
    
    # Get numbers at start and end of collection period
    start_exchanges = updates[updates['timebank'] == df['timebank'][row]]['exchanges'].min()
    end_exchanges = updates[updates['timebank'] == df['timebank'][row]]['exchanges'].max()
    
    start_hours = updates[updates['timebank'] == df['timebank'][row]]['hours'].min()
    end_hours = updates[updates['timebank'] == df['timebank'][row]]['hours'].max()
    
    # Divide difference by number of days
    df['avg_daily_exchanges'][row] = (end_exchanges - start_exchanges) / total_days
    df['avg_daily_hours'][row] = (end_hours - start_hours) / total_days
    
    # Get average hours per exchange
    if df['avg_daily_exchanges'][row] == 0:
        df['hours_per_exchange'][row] = 0
    else:
        df['hours_per_exchange'][row] = df['avg_daily_hours'][row] / df['avg_daily_exchanges'][row]

# Get average offer:request ratio and append to df
df['avg_offers'] = ''
df['avg_requests'] = ''
df['offer_request_ratio'] = ''

for row in range(len(df)):
    offers = updates[updates['timebank'] == df['timebank'][row]]['offers'].mean()
    requests = updates[updates['timebank'] == df['timebank'][row]]['requests'].mean()
    
    df['avg_offers'][row] = offers
    df['avg_requests'][row] = requests
    
    try:
        df['offer_request_ratio'][row] = offers / requests
    except:
        df['offer_request_ratio'][row] = 0

df['members_starting'] = ''
df['members_daily_new'] = ''

for row in range(len(df)):
    
    # Get member numbers at start and end of collection period
    start_members = updates[updates['timebank'] == df['timebank'][row]]['members'].iloc[0]
    end_members = updates[updates['timebank'] == df['timebank'][row]]['members'].iloc[-1]
    
    df['members_starting'][row] = start_members
    
    # Get average daily new members
    df['members_daily_new'][row] = (end_members - start_members) / total_days

df.head(15)

### Add info from `social.csv`

social = open_data('social')
social.head(3)

# Replace missing values with 0
social.replace([np.nan, 'page unavailable', 'account suspended'], 0, inplace=True)

# Append social columns to df
df = pd.merge(df, social[['timebank', 'facebook_likes', 'facebook_followers',
                    'twitter_tweets', 'twitter_following', 'twitter_followers']], on='timebank')

### Add category info from `talents.csv`, `offers.csv`, and `requests.csv`

# Open category files and change timestamp to day only
talents = open_data('talents')
talents['timestamp'] = talents['timestamp'].dt.floor('D')

offers = open_data('offers')
offers['timestamp'] = offers['timestamp'].dt.floor('D')

requests = open_data('requests')
requests['timestamp'] = requests['timestamp'].dt.floor('D')

# Merge all into categories dataframe
merge_columns = ['cat_id', 'cat_parent', 'category', 'timebank', 'timestamp']
categories = talents.merge(offers, how='left', on=merge_columns)
categories = categories.merge(requests, how='left', on=merge_columns)

# Replace nan values with 0
categories.replace(np.nan, 0, inplace=True)

# Reorder columns for viewing
categories = categories[['timestamp', 'timebank', 'category', 'cat_id', 'cat_parent',
           'count_talent', 'count_offers', 'count_requests']]

# Get values of offers - requests
categories['offers_requests_diff'] = categories['count_offers'] - categories['count_requests']

# Save to .csv
categories.to_csv('../data/categories.csv', index=False)

df['categories_total'] = ''
df['categories_parent'] = ''
df['categories_with_offers'] = ''
df['categories_with_requests'] = ''

for row in range(len(df)):
    
    # Filter to specific timebank
    tb = categories[categories['timebank'] == df['timebank'][row]]

    # Get total number of categories
    total_cat = tb['cat_id'].nunique()
    df['categories_total'][row] = total_cat

    # Get number of parent categories
    df['categories_parent'][row] = tb[tb['cat_parent'] == 'is_parent']['cat_id'].nunique()
    
    # Get percent of categories with offers
    try:
        df['categories_with_offers'][row] = tb[tb['count_offers'] != 0]['cat_id'].nunique() / total_cat
    except:
        df['categories_with_offers'][row] = 0
    
    # Get percent of categories with requests
    try:
        df['categories_with_requests'][row] = tb[tb['count_requests'] != 0]['cat_id'].nunique() / total_cat
    except:
        df['categories_with_requests'][row] = 0

df['talent_per_cat_per_member'] = ''

for row in range(len(df)):
    
    # Filter to specific timebank
    tb = categories[categories['timebank'] == df['timebank'][row]]
    
    # Get average total talents for all parent categories
    catparent = tb[tb['cat_parent'] == 'is_parent']['count_talent'].groupby(tb['timestamp']).sum().mean()
    
    # Get total talents / number of parent categories / total members
    try:
        df['talent_per_cat_per_member'][row] = catparent / df['categories_parent'][row] / df['members_starting'][row]
    except:
        df['talent_per_cat_per_member'][row] = 0

# # Look at specific timebank
# categories[categories['timebank'] == 'addington'].sort_values('offers_requests_diff')

### Convert all numerical columns to `float`

df.loc[:, df.columns != 'timebank'] = df.loc[:, df.columns != 'timebank'].astype(float)

# Preview
df.head(10).T

### Export to `combined.csv`

# Include dates covered in file name
start_date = updates['timestamp'].min().strftime('%Y-%m-%d')
end_date = updates['timestamp'].max().strftime('%Y-%m-%d')

df.to_csv(f'../data/combined_{start_date}_{end_date}.csv', index=False)
